リファクタリング: モジュール構造整理と検索機能改善
- core/, discord/, infra/, llm/, search/ディレクトリに機能分割
- SearchContextResult導入でステータス詳細化
- NO_RESULTSケースの明確化
- ログ出力の改善
