+++ app/src/sub/core/__init__.py
@@ -1,0 +1,1 @@
+"""Core domain (dataclasses, base config structures)."""
