+++ app/src/sub/discord/__init__.py
@@ -1,0 +1,1 @@
+"""Discord integration helpers (message conversion, channel/thread utils)."""
