+++ app/src/sub/infra/__init__.py
@@ -1,0 +1,1 @@
+"""Infrastructure utilities (logging, rate limiting, dedup, cache helpers)."""
