+++ app/src/sub/search/__init__.py
@@ -1,0 +1,1 @@
+"""Web search orchestration layers (decision, execution, formatting)."""
